v2.13.9
fix: resolve 18 bugs across connectors, crypto, and protocol layers

- aes: replace non-thread-safe global-state implementation with
  OpenSSL EVP layer; eliminates data races under concurrent use

- recordset: fix inverted null checks in get_long/get_double/get_string
  (was throwing on valid values, passing on nulls)

- config: fix false positive error on integer array value of 0;
  now bounds-checks via config_setting_length before reading

- hdfs: fix hdfsRead() called with &buffer instead of buffer in read();
  correct inverted return logic (was returning 0 on success);
  fix list() filter checking uninitialized file.name instead of filename

- s3: fix out-of-bounds write in chwd() path normalization
  (assignment-in-condition was UB; changed to comparison)

- oracle: fix impossible double condition in resultset::next()
  (OCI_SUCCESS && OCI_NO_DATA can never both be true; changed to ||);
  fix get_string() using != instead of == for SQLT type comparisons

- dmi: add null check on addrp before dereference in address_from_efi()

- socket: change sent/remain to ssize_t in write(); skip remain update
  on send() error to prevent integer wrap and infinite loop

- rabbit: fix consumer::get() early-return condition (&& -> ||)

- mml: make retval/running/looping/finished atomic to eliminate
  data race between exec() and its timeout thread; remove by-ref
  short int from callback signature to match atomic usage

- ldap: set _port when non-zero port is passed to connect()

- file: zero-initialize szr and szw in copy() to prevent UB reads

- kinesis: guard _consumer->detach() with null check in destructor;
  use this->th explicitly in lambda to reference member vector

- samba: increase put() buffer from hardcoded 1024 to MAX_BUFFER_SIZE

- connection/file/ftp/hdfs/s3/samba/ssh/nothing headers + impls:
  change list() return type from int to size_t; change copy/get/put
  return types from long long to off_t throughout all connectors

version: 2.13.8-7988 -> 2.13.9-8029

+++ include/axon/connection.h
@@ -139,17 +139,17 @@ namespace axon
 				virtual bool chwd(std::string) = 0;
 				virtual std::string pwd() = 0;
 				virtual bool mkdir(std::string) = 0;
-				virtual int list(const cb &) = 0;
-				virtual int list(std::vector<axon::entry> &) = 0;
-				virtual long long copy(std::string, std::string, bool) = 0;
-				virtual long long copy(std::string, std::string) = 0;
+				virtual size_t list(const cb &) = 0;
+				virtual size_t list(std::vector<axon::entry> &) = 0;
+				virtual off_t copy(std::string, std::string, bool) = 0;
+				virtual off_t copy(std::string, std::string) = 0;
 				virtual bool ren(std::string, std::string) = 0;
 				virtual bool del(std::string) = 0;
 
-				virtual long long get(std::string, std::string, bool) = 0;
-				virtual long long put(std::string, std::string, bool) = 0;
-				virtual long long put(std::string src, std::string dest) final { return put(src, dest, false); };
-				virtual long long put(std::string src) final { return put(src, src, false); };
+				virtual off_t get(std::string, std::string, bool) = 0;
+				virtual off_t put(std::string, std::string, bool) = 0;
+				virtual off_t put(std::string src, std::string dest) final { return put(src, dest, false); };
+				virtual off_t put(std::string src) final { return put(src, src, false); };
 
 				virtual bool open(std::string, std::ios_base::openmode) = 0;
 				virtual bool close() = 0;
