+++ src/nothing.cpp
@@ -21,17 +21,17 @@ namespace axon
 		bool nothing::chwd(std::string) { return true; };
 		std::string nothing::pwd()  { return "/"; };
 		bool nothing::mkdir(std::string) { return true; };
-		int nothing::list(const axon::transfer::cb &) { return 0; };
-		int nothing::list(std::vector<entry> &) { return 0; };
-		long long nothing::copy(std::string , std::string , bool ) { return 0L; };
-		long long nothing::copy(std::string src, std::string dest) { return copy(src, dest, false); };
+		size_t nothing::list(const axon::transfer::cb &) { return 0; };
+		size_t nothing::list(std::vector<entry> &) { return 0; };
+		off_t nothing::copy(std::string , std::string , bool ) { return 0L; };
+		off_t nothing::copy(std::string src, std::string dest) { return copy(src, dest, false); };
 		bool nothing::ren(std::string , std::string ) { return true; };
 		bool nothing::del(std::string ) { return true; };
 
 		int nothing::cb(const struct entry *) { return 0; };
 
-		long long nothing::get(std::string , std::string , bool ) { return 0L; };
-		long long nothing::put(std::string , std::string , bool ) { return 0L; };
+		off_t nothing::get(std::string , std::string , bool ) { return 0L; };
+		off_t nothing::put(std::string , std::string , bool ) { return 0L; };
 
 		bool nothing::open(std::string, std::ios_base::openmode) { return true; }
 		bool nothing::close() { return true; }
